requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
python-dotenv==1.2.1
requests==2.32.5
SQLAlchemy==2.0.44
urllib3==2.5.0
Werkzeug==3.0.4
